change variables of type int to size_t

+++ tasks/associative_operation/associative_operation.cpp
@@ -1,13 +1,13 @@
 #include "associative_operation.h"
 
 bool IsAssociative(const std::vector<std::vector<size_t>>& table) {
-    int n = table.size();
-    for (int i = 0; i < n; ++i) {
+    size_t n = table.size();
+    for (size_t i = 0; i < n; ++i) {
         if (table[i].size() != n) {
             return false;
         }
-        for (int j = 0; j < n; ++j) {
-            for (int k = 0; k < n; ++k) {
+        for (size_t j = 0; j < n; ++j) {
+            for (size_t k = 0; k < n; ++k) {
                 if (table[table[i][j]][k] != table[i][table[j][k]]) {
                     return false;
                 }
